Financial-AI-Analyst/requirements.txt
phidata
yfinance
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
requests